Handle snappy & gzip compression
Support for handling compression in both consumer and producer.
On the consumer side it's handled transparently, for producers
there's a new property in the payload (`attributes`) that can be
set to 0, 1, or 2 (defaults to 0).

This PR also adds support for consumer.encoding which can be used
to control how values are returned. Previously it was only possible
to retrieve a stringified version of the value which destroyed
binary encodings (like msgpack). The default is "utf8" which should
match the previous behavior.

Fixes https://github.com/SOHU-Co/kafka-node/issues/27