Buffer batch for async producers
Other kafka connectors such as the Java one have the possibility to configurate an optional batch (client-side) for async producer mode, in order to send batches of messages instead of single messages at time. This increases performance because the kafka server doesnt need to answer ACK at every message.

The options are:
- Batch size in bytes
- Batch age in miliseconds