Kafka topics should not contain special characters (#492)
* Kafka topics should not contain special characters

This is my observation while playing with createTopics function. I later realised that topic names which I am creating contains special characters such as :,// etc which is causing the issue. I tried without them and it works. It does work with other number and string combinations so no test required for that.

Hope this PR is help.

* Update client.js

Fixes for build failure #1.

* Update client.js

Fixed for build #2.

* Update client.js

Fixes for build #3.

* Validating the topics using util function

* Build fix

* Update client.js

Updating the name of function as per changes in pull req no 17.

* Handling async nature of validation.

Build fix.

* Update client.js

Build fix.

* Update client.js

For build fix.