Creating a writable commit stream.
The commit stream is a writable stream that you can pipe events to and
it will commit them. This is useful so that you can pipe the
ConsumerStream through a transform stream and be completely confident
that the only way you will commit an offset is by successfully
processing the message.