auto commit and add event for rebalance
Should resolve #787

If auto commit is true, commit before we rejoin the group on a rebalance
According to the kafka protcol, commits are allowed during this stage, as the generation ID has not been bumped yet.

If auto commit is disabled, implementors need a way to decide what to commit, so an onRebalance option was added.

Implementors can perform commits during this method and call back when done to finish the rebalance.