Add create topics support
Adds support for sending a create topics request as a real protocol request. This supports specifying number of partitions and replication factor.
Explicit replica assignments and config entries are left out for now.

A create topic request must be sent to the controller. To find the current controller, this also adds a version 1 metadata request, which includes the controller Id in the response.