Add create topics support (#958)
* Add create topics support

Adds support for sending a create topics request as a real protocol request. This supports specifying number of partitions and replication factor.
Explicit replica assignments and config entries are left out for now.

A create topic request must be sent to the controller. To find the current controller, this also adds a version 1 metadata request, which includes the controller Id in the response.

* Handle different error message in kafka versions

Error message changes a bit depending on version. This relaxes the assertion a bit.

* Remove metadata api support version

metadata v1 is not supported until 0.10

* Return error if controller cannot be found

Happens when cluster is is running version less than 0.10

* Parse partition metadata correctly for v1

The order resulted in isInternal being used as the number of partitions during decoding.

* Address PR feedback

- Add createTopics alias on admin client
- Remove V2 suffix, replacing old versions of createTopics and loadMetadataForTopics on KafkaClient
- Add loadMetadata function
- Use supported API to resolve metadata request version

* Encode null value properly

* Forward old createTopics calls to previous impl

* Update README with new createTopics function

* Update cached metadata in getController

* Remove unused sandbox

* Add additional assertion of createTopic test

Now fetches metadata again to verify topics were created with requested partition count and replication factor.

* Add sendControllerRequest method

Wraps the provided callback with retry logic if NotController error was returned.
In this case the cached controller is cleared and it is fetched again by getting the latest metadata.