Add config entries and replica assignment to create topic
Adds support for providing config entries and explicit replica assignment
when creating new topics.

The existing create topic implementation has been cleaned up a bit,
as it didn't make use of the existing API support setup (it was fixed to protocol version 1).
As part of the cleanup, version 0 and 1 of the request is now supported.

The send controller request was also cleaned up a bit. A new function sendRequestToBroker
provides a more robust way to send requests to a specific broker now.