Fix ConsumerGroupStream commit payload type
Currently, the payload argument for the `commit`
method is declared as type `Message` interface which
in turn expects a `value` field that offset requests do
not have (This forces workarounds like inserting a dummy
value to get around the type-checker).

This patch fixes this by specifying the commit payload
to be of type `OffsetCommitRequest` which declares
the sufficient fields of commit payloads.