Defaults createTopics' param async=false
Async must default to false.
It is dangerous to use async=true for createTopics unless explicitly specified.

For running async calls, the subsequent steps (such as addTopics) will fail as the topics are yet to be created/in the progress of creation.
It is the nature of async to return immediately even before the createTopics is completed.

This fixes the regression from the fix 5fd3764ad8a2628e675bc05c4b6268edffd0dc26.